// src/@types/enums/cities.ts
export enum Cities {
  PORTO_ALEGRE = "Porto Alegre, RS",
  RIO_DE_JANEIRO = "Rio de Janeiro, RJ",
  CURITIBA = "Curitiba, PR",
  SAO_PAULO = "São Paulo, SP",
}

export enum CitiesWithoutStates {
  PORTO_ALEGRE = "Porto Alegre",
  RIO_DE_JANEIRO = "Rio de Janeiro",
  CURITIBA = "Curitiba",
  SAO_PAULO = "São Paulo",
}